Add --frozen flags and use trunk in offline mode
This should prevent the Cargo.lock file from being automatically
adjusted, causing "uncommitted changes" warnings in the info popup.

Using trunk in offline mode should prevent it from downloading a broken
version of tooling on aarch64, causing build failures on my server which
require changes in the Dockerfile and building the frontend separately
from the server.
Offline mode in trunk v0.18.7 was broken, never recognising the wasm-opt
binary provided by the system due to a bug. This was fixed in v0.18.8
which has just been pushed to the alpine edge repositories.