Improve server error handling on the client
The server will now send json-encoded SerializableErrors if the client
explicitly requests `application/json` as one of the accepted formats,
which the frontend will now do.

This allows the client to display the server error stack in its own
error stacks, which is more readable than a double "caused by".

also add some documentation in the error_handling crate