Send channel fetching progress to clients
Search by channel endpoints will no longer hang until the channel is
indexed, as that was causing timeouts on large channels.

Instead, the indexing will be scheduled as a background job, and the
server will wait at most 1 second for it to complete when a new request
comes in.
If it does not complete within that timeframe, the server will send a
custom "333 Not Ready Yet" response with progress encoded as json.
Clients should request the same URL again when they receive a '333'
response. Clients may decode the json payload and display the progress
status to the user.